refactor: move container back to pages
