+++ src/server/pages/members.ts
@@ -14,8 +14,7 @@ export function getMembersPageData(members: Members): MembersPageData {
     throw new Error('Please convert Members data at first.');
   }
 
-  const resultData = members.result;
-  return resultData.map(member => ({
+  return members.result.map(member => ({
     name: member.name,
     nameNotations: member.nameNotations,
     join: member.join,
